+++ PF-Front/src/components/Deshboard/Admin/Admin.jsx
@@ -6,11 +6,7 @@ import AdminBookings from "./AdminBookings/AdminBookings";
 import AdminUsers from "./AdminUsers/AdminUsers";
 // import AdminAdmins from "./AdminAdmins/AdminAdmins";
 
-
-
 const Admin = () => {
-
-
     const [selectedTab, setSelectedTab] = useState(0);
 
     const handleTabChange = (event, newValue) => {
@@ -28,11 +24,11 @@ const Admin = () => {
     });
 
     return (
-
-        <Grid container justifyContent="center"
-            sx={{ width: 'auto', backgroundColor: '#FAFAFF' }}>
-
-            <div style={{ width: '80%', margin: '20px' }}>
+        <Grid
+            container
+            justifyContent="center"
+            sx={{ width: "auto", backgroundColor: "#FAFAFF" }}>
+            <div style={{ width: "80%", margin: "20px" }}>
                 <Typography
                     variant="h1"
                     sx={{
@@ -40,9 +36,8 @@ const Admin = () => {
                         fontWeight: "bold",
                         color: "#0400CB",
                         textAlign: "Start",
-                        margin: '20px'
-                    }}
-                >
+                        margin: "20px",
+                    }}>
                     DASHBOARD
                 </Typography>
                 <Tabs
@@ -62,7 +57,7 @@ const Admin = () => {
                         }}
                     />
                     <Tab
-                        label="Create Rooms"
+                        label="Manage Rooms"
                         sx={{
                             backgroundColor: "#F3F3F7",
                             color: "#868688",
@@ -97,7 +92,6 @@ const Admin = () => {
 
                     <Tab
                         label="Custome Admins"
-
                         sx={{
                             backgroundColor: "#F3F3F7",
                             color: "#868688",
@@ -106,7 +100,7 @@ const Admin = () => {
                                 backgroundColor: "#EFEEFF",
                             },
                         }}
-                    /> 
+                    />
                 </Tabs>
                 {selectedTab === 0 && <AdminHotel />}
                 {selectedTab === 1 && <AdminRooms />}
@@ -114,7 +108,6 @@ const Admin = () => {
                 {selectedTab === 3 && <AdminUsers />}
                 {selectedTab === 4 && <AdminAdmins />}
             </div>
-
         </Grid>
     );
 };
